refactor(analysis): extract complexity risk helpers
